// src/types.ts
export type ChecklistStatus = 'yes' | 'no' | 'pending';

export interface ChecklistMap {
  [date: string]: ChecklistStatus[];
}
// ...
export interface UserProfile {
  name: string;
  jobTitle: string;
  role: 'admin' | 'hod' | 'assistant_manager' | 'officer' | 'manager' | 'sr_executive' | 'executive';
  tierId: string;
  email: string;
  employeeId?: string;
  assignedUnit?: string;
  shift?: string;
  photoURL?: string;
  googleUid?: string;
}